# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Bolsa de Trabajo",
        "description": "Documentación de la API desarrollada en Laravel para la bolsa de trabajo.",
        "contact": {
            "name": "Soporte",
            "email": "soporte@tuapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor Laravel local"
        }
    ],
    "paths": {
...
        "/api/postulaciones": {
            "get": {
                "tags": [
                    "Postulaciones"
                ],
                "summary": "Obtener todas las postulaciones del estudiante",
                "operationId": "getPostulaciones",
                "responses": {
                    "200": {
                        "description": "Lista de postulaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "postulaciones": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Postulaciones obtenidas con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Postulaciones"
                ],
                "summary": "Crear una nueva postulación",
                "operationId": "postular",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Id_Trabajo"
                                ],
                                "properties": {
                                    "Id_Trabajo": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Postulación creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "postulacion": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/postulaciones/curriculum": {
            "get": {
                "tags": [
                    "Postulaciones"
                ],
                "summary": "Obtener el currículum en PDF del estudiante",
                "operationId": "getCurriculumPDF",
                "responses": {
                    "200": {
                        "description": "Currículum encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cv": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Currículum no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}